routes refactor
